Steamlit-superceded/requirements.txt
pandas 
plotly
streamlit_option_menu
opencv-python-headless
